// domain/snipe/enums/Snipe.enum.ts
export enum ESnipeRankingSort {
    Count = "Count",
    AveragePP = "AveragePP",
    AverageStars = "AverageStars",
    WeightedPP = "WeightedPP",
}

export enum ESnipePlayerListSort {
    PP = "PP",
    Accuracy = "Accuracy",
    Stars = "Stars",
    Misses = "Misses",
    Date = "Date",
}

export enum ESnipePlayerChangeType {
    Gain = "Gain",
    Loss = "Loss",
}